Added and Updated some things (See explanation)
Added basic Player 2 functionality

Updated debug mode: Direct Win/Lose Screen from lvl1

Fixed Stage Clear

Fixed Conversion int <-> float that might leak some memory

Fixed Shoot in Main Menu

Fixed Camera Limits

+++ Last_Resort/ModuleSceneLvl1.h
@@ -41,9 +41,9 @@ public:
 
 public: //movement of the firstplanebackground
 	
-	int FirstPlaneBackGround_position_X = 0;
-	int SecondPlaneGround_position_X = 0;
-	int ThirdPlaneBackground_position_X = 0;
+	float FirstPlaneBackGround_position_X = 0;
+	float SecondPlaneGround_position_X = 0;
+	float ThirdPlaneBackground_position_X = 0;
 
 	float ScrollingSpeed_Foreground=1.0f;
 };
